requirements.txt
alpaca-py==0.40.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
exceptiongroup==1.3.0
exchange_calendars==4.5.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
korean-lunar-calendar==0.3.1
msgpack==1.1.0
numpy==2.0.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pandas_market_calendars==5.1.0
pydantic==2.11.4
pydantic_core==2.33.2
pyluach==2.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sseclient-py==1.8.0
tastytrade==10.2.1
toolz==1.0.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
